♻️ refactor: reservation context api로 리팩토링
