+++ src/contexts/ReservationProvider/context.ts
@@ -1,0 +1,22 @@
+import { createContext } from "react";
+
+interface ContextProps {
+  reservation: any;
+  handleInitReservation: any;
+  handleSetCurrentBlock: any;
+  handleClose: any;
+  handleStartCreate: any;
+  handleStartUpdate: any;
+  handleDecreaseStep: any;
+  handleCreateReservation: any;
+  handleUpdateReservation: any;
+  handleDeleteReservation: any;
+  handleChangeHasBall: any;
+  handleClickReservationMarker: any;
+  handleSetCurrentInput: any;
+  handleSetTime: any;
+}
+
+const ReservationContext = createContext<ContextProps>({} as ContextProps);
+
+export default ReservationContext;
