+++ src/contexts/ReservationProvider/hook.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import ReservationContext from "./context";
+
+const useReservationContext = () => useContext(ReservationContext);
+
+export default useReservationContext;
