+++ src/contexts/hooks.ts
@@ -2,3 +2,4 @@ export { default as useAuthContext } from "@contexts/AuthProvider/hook";
 export { default as useNavigationContext } from "@contexts/NavigationProvider/hook";
 export { default as useMapContext } from "@contexts/MapProvider/hook";
 export { default as useSocketContext } from "@contexts/SocketProvider/hook";
+export { default as useReservationContext } from "@contexts/ReservationProvider/hook";
