+++ src/contexts/index.ts
@@ -2,3 +2,4 @@ export { default as AuthProvider } from "./AuthProvider";
 export { default as NavigationProvider } from "./NavigationProvider";
 export { default as MapProvider } from "./MapProvider";
 export { default as SocketProvider } from "./SocketProvider";
+export { default as ReservationProvider } from "./ReservationProvider";
